enkel klant verwijderen indien geen facturen + ordernummer toegevoegd + verkoopsvoorwaarden afdrukken

+++ src/main/java/be/boomkwekerij/plant/controller/CustomerController.java
@@ -1,8 +1,11 @@
 package be.boomkwekerij.plant.controller;
 
 import be.boomkwekerij.plant.model.dto.CustomerDTO;
+import be.boomkwekerij.plant.model.dto.InvoiceDTO;
 import be.boomkwekerij.plant.service.CustomerService;
 import be.boomkwekerij.plant.service.CustomerServiceImpl;
+import be.boomkwekerij.plant.service.InvoiceService;
+import be.boomkwekerij.plant.service.InvoiceServiceImpl;
 import be.boomkwekerij.plant.util.CrudsResult;
 import be.boomkwekerij.plant.util.ExceptionUtil;
 import be.boomkwekerij.plant.util.SearchResult;
@@ -10,6 +13,7 @@ import be.boomkwekerij.plant.util.SearchResult;
 public class CustomerController {
 
     private CustomerService customerService = new CustomerServiceImpl();
+    private InvoiceService invoiceService = new InvoiceServiceImpl();
 
     public CrudsResult createCustomer(CustomerDTO customerDTO) {
         try {
@@ -45,10 +49,24 @@ public class CustomerController {
 
     public CrudsResult deleteCustomer(String id) {
         try {
+            if (hasInvoices(id)) {
+                return createCrudsError("Er zijn nog facturen gelinkt aan deze klant!");
+            }
             return customerService.deleteCustomer(id);
         } catch (Exception e) {
             return createCrudsError(e);
         }
+    }
+
+    private boolean hasInvoices(String id) {
+        boolean hasInvoices = false;
+        SearchResult<InvoiceDTO> allInvoicesFromCustomer = invoiceService.getAllInvoicesFromCustomer(id);
+        if (allInvoicesFromCustomer.isSuccess()) {
+            if (allInvoicesFromCustomer.getResults().size() > 0) {
+                hasInvoices = true;
+            }
+        }
+        return hasInvoices;
     }
 
     public SearchResult<CustomerDTO> getAllCustomersWithName(String name) {
@@ -63,6 +81,13 @@ public class CustomerController {
         SearchResult<CustomerDTO> failure = new SearchResult<CustomerDTO>();
         failure.setSuccess(false);
         failure.getMessages().add(e.getMessage());
+        return failure;
+    }
+
+    private CrudsResult createCrudsError(String message) {
+        CrudsResult failure = new CrudsResult();
+        failure.setSuccess(false);
+        failure.getMessages().add(message);
         return failure;
     }
